requirements.txt
fastapi
uvicorn[standard]
apscheduler
aiosqlite
httpx
feedparser
yfinance
beautifulsoup4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyyaml
python-dotenv
